+++ addon/components/paper-card-content.js
@@ -1,7 +1,7 @@
 import Ember from 'ember';
-import FlexMixin from '../mixins/flex-mixin';
 
-export default Ember.Component.extend(FlexMixin, {
-  tagName: 'md-card-content',
-  classNames: ['paper-card-content']
+const { Component } = Ember;
+
+export default Component.extend({
+  tagName: 'md-card-content'
 });