+++ addon/components/paper-card-image.js
@@ -1,0 +1,9 @@
+import Ember from 'ember';
+
+const { Component } = Ember;
+
+export default Component.extend({
+  tagName: 'img',
+  classNames: ['md-card-image'],
+  attributeBindings: ['src', 'title', 'alt']
+});