+++ blueprints/ember-paper/index.js
@@ -1,16 +1,16 @@
 'use strict';
 
 module.exports = {
-  normalizeEntityName() {
+  normalizeEntityName: function() {
     // this prevents an error when the entityName is
     // not specified (since that doesn't actually matter
     // to us
   },
 
-  afterInstall() {
+  afterInstall: function() {
     var _this = this;
     return this.addBowerPackagesToProject([
-      {name: 'hammerjs', target:'latest'},
+      {name: 'hammer.js', target:'latest'},
       {name: 'matchMedia', target: '0.2.0'}
     ]).then(function() {
       return _this.addPackagesToProject([