+++ tests/dummy/app/helpers/is-subheading.js
@@ -1,0 +1,8 @@
+import Ember from 'ember';
+const { Helper, typeOf } = Ember;
+
+export default Helper.extend({
+  compute([row]) {
+    return typeOf(row) !== 'array';
+  }
+});